Fix bug, and make curl_error
The bug was in the file:include/_curl.h. In this file, the class "curl_easy" has some pomblie in the Constructor, if you try to using this by mutli, it will try to mutliple uses curl_global_init.
To fix this bug, I add the static mamber: "_global_init" to record wheather it is initialized or not, and add static function: "global_init()" in order to facilitate global initialization.

The error mechanism is implemented using exceptions(for now, more will be introduced later).